Siirrä getAineistoSchema ja getAineistotSchema schemas/common:n alle
