+++ src/schemas/common.ts
@@ -3,6 +3,14 @@ import { IlmoitettavaViranomainen } from "@services/api";
 import filter from "lodash/filter";
 import { yhteystietoSchema } from "./yhteystieto";
 
+const getAineistoSchema = () =>
+  Yup.object().shape({
+    dokumenttiOid: Yup.string().required(),
+    nimi: Yup.string().required(),
+    jarjestys: Yup.number().integer().notRequired(),
+  });
+export const getAineistotSchema = () => Yup.array().of(getAineistoSchema()).nullable();
+
 export const ilmoituksenVastaanottajat = () =>
   Yup.object()
     .shape({
